+++ CHANGELOG.md
@@ -4,6 +4,35 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+
+### Vsix
+
+
+### VB -> C#
+* Cast foreach collection to IEnumerable if needed or unknown
+* Fix ordering bug converting redim bounds without preserve for 2d arrays
+* Exit Property should become return [#497](https://github.com/icsharpcode/CodeConverter/issues/497)
+* First effort converting some Xml Member Access
+* Avoid adding new keyword when not allowed/required [#504](https://github.com/icsharpcode/CodeConverter/issues/504)
+
+### C# -> VB
+
+
+## [7.6.0] - 2020-01-11
+
+
+### Vsix
+
+* Fix Visual Studio crash when converting some structures
+
+### VB -> C#
+
+
+### C# -> VB
+
+
+## [7.5.0] - 2020-01-11
+
 * Improve ambiguous name resolution
 
 ### Vsix